Try to move getPeaks and allow to getSimulatedSpectrum directly from data/examples (using Lorentzian)
