+++ __tests__/data/examples.js
@@ -25,8 +25,39 @@ let simple4 = new Chromatogram(
 );
 
 
+// https://en.wikipedia.org/wiki/Cauchy_distribution
+function lorentzian(x, x0 = 0, gamma = 1) {
+    return (gamma * gamma) / (Math.PI * gamma * (gamma * gamma + (x - x0) * (x - x0)));
+}
+
+function getSimulatedSpectrum(options) {
+    const {
+        size=30
+    } = options;
+    const fourth = size / 4;
+    let times = new Array(size);
+    let tic = new Array(size);
+    let ms = new Array(size);
+    for (let i = 0; i < size; ++i) {
+        times[i] = i;
+        tic[i] = lorentzian(i, fourth) + 2 * lorentzian(i, 2 * fourth) + lorentzian(i, 3 * fourth);
+        ms[i] = [[1, 2, 3], [1, 1, 1]];
+    }
+    let chromatogram = new Chromatogram(times,
+        {
+            'tic': tic,
+            'ms': ms
+        }
+    );
+
+    return chromatogram;
+};
+
+
+
 module.exports = {
     simple,
     simpleStringified,
-    simple4
+    simple4,
+    getSimulatedSpectrum
 };
